Move preview repository comparison
